+++ CHANGELOG.md
@@ -10,11 +10,13 @@ and this project adheres to
 
 - Update soroban-sdk version from v20.0.3 to v20.1.0 ([#193])
 - Fixes documentation and naming ([#200])
+- Multihop: adds a new field in the Swap struct, that hold max_belief_price ([234])
 - Fixes incorrect assignment of total_fee_bps in both pool and pool stable ([235])
 - Pool: adds a missed part of return_amount argument ([#238])
 - Pool: Replace panic! with panic_with_error! to provide more contextual information ([#206])
 
 [#200]: https://github.com/Phoenix-Protocol-Group/phoenix-contracts/pull/200
+[#234]: https://github.com/Phoenix-Protocol-Group/phoenix-contracts/pull/234
 [#235]: https://github.com/Phoenix-Protocol-Group/phoenix-contracts/pull/235
 [#238]: https://github.com/Phoenix-Protocol-Group/phoenix-contracts/pull/238
 [#206]: https://github.com/Phoenix-Protocol-Group/phoenix-contracts/pull/206